# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  service.CreateClusterRequest:
    properties:
      alias:
        type: string
      apiEndpoint:
        description: 移除 required，因为 KubeConfig 中已经包含
        type: string
      description:
        type: string
      kubeConfig:
        type: string
      name:
        type: string
      provider:
        type: string
      region:
        type: string
      userId:
        type: integer
    required:
    - kubeConfig
    - name
    type: object
  service.UpdateClusterRequest:
    properties:
      alias:
        type: string
      apiEndpoint:
        type: string
      description:
        type: string
      kubeConfig:
        type: string
      name:
        type: string
      provider:
        type: string
      region:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 运维管理平台 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: iom API
  version: "1.0"
paths:
...
  /plugins/kubernetes/clusters:
    get:
      consumes:
      - application/json
      description: 获取所有 Kubernetes 集群
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      summary: 获取集群列表
      tags:
      - Kubernetes/Cluster
    post:
      consumes:
      - application/json
      description: 创建新的 Kubernetes 集群
      parameters:
      - description: 集群信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.CreateClusterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 创建集群
      tags:
      - Kubernetes/集群管理
  /plugins/kubernetes/clusters/{id}:
    delete:
      consumes:
      - application/json
      description: 删除 Kubernetes 集群
      parameters:
      - description: 集群ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 删除集群
      tags:
      - Kubernetes/集群管理
    get:
      consumes:
      - application/json
      description: 获取 Kubernetes 集群详细信息
      parameters:
      - description: 集群ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      summary: 获取集群详情
      tags:
      - Kubernetes/Cluster
    put:
      consumes:
      - application/json
      description: 更新 Kubernetes 集群信息
      parameters:
      - description: 集群ID
        in: path
        name: id
        required: true
        type: integer
      - description: 集群信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.UpdateClusterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 更新集群
      tags:
      - Kubernetes/集群管理
  /plugins/kubernetes/clusters/{id}/config:
    get:
      consumes:
      - application/json
      description: 获取集群的 KubeConfig 配置（已解密）
      parameters:
      - description: 集群ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      summary: 获取集群凭证
      tags:
      - Kubernetes/Cluster
  /plugins/kubernetes/clusters/{id}/sync:
    post:
      consumes:
      - application/json
      description: 同步指定集群的状态信息（节点数、Pod数等）
      parameters:
      - description: 集群ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      summary: 同步集群状态
      tags:
      - Kubernetes/Cluster
  /plugins/kubernetes/clusters/{id}/test:
    post:
      consumes:
      - application/json
      description: 测试 Kubernetes 集群连接状态
      parameters:
      - description: 集群ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      summary: 测试集群连接
      tags:
      - Kubernetes/Cluster
...
  /plugins/kubernetes/clusters/kubeconfig/existing:
    get:
      consumes:
      - application/json
      description: 获取当前用户在指定集群的最新KubeConfig
      parameters:
      - description: 集群ID
        in: query
        name: clusterId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 获取用户现有的KubeConfig
      tags:
      - Kubernetes/Cluster
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"